# swagger.yaml
swagger: "2.0"
info:
  title: "API de Predição de Iris"
  description: "API para autenticação e predição usando um modelo de machine learning"
  version: "1.0.0"
servers:
  - url: http://localhost:5001
    description: Local server

securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header

# aplica JWT por padrão a todas as operações;
# no login, sobrescrevemos com security: []
security:
  - JWT: []

paths:
  /login:
    post:
      summary: "Endpoint para autenticação de usuário"
      security: []   # não exige token
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              username:
                type: string
                example: admin
              password:
                type: string
                example: secret
      responses:
        200:
          description: Token JWT gerado com sucesso.
          schema:
            type: object
            properties:
              token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        401:
          description: Credenciais inválidas.
          
  /predict:
    post:
      summary: "Endpoint protegido por token para obter predição"
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              sepal_length:
                type: number
                format: float
                example: 5.1
              sepal_width:
                type: number
                format: float
                example: 3.5
              petal_length:
                type: number
                format: float
                example: 1.4
              petal_width:
                type: number
                format: float
                example: 0.2
      responses:
        200:
          description: Predição realizada com sucesso.
          schema:
            type: object
            properties:
              predicted_class:
                type: integer
                example: 0
        400:
          description: Dados inválidos.
        401:
          description: Token inválido ou ausente.

  /predictions:
    get:
      parameters:
        - name: limit
          in: query
          type: integer
          description: Quantidade de registros a serem retornados (padrão = 10).
          example: 5
        - name: offset
          in: query
          type: integer
          description: A partir de qual registro começar (padrão = 0).
          example: 10
      responses:
        200:
          description: Lista de predições.
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                sepal_length:
                  type: number
                  format: float
                  example: 5.1
                sepal_width:
                  type: number
                  format: float
                  example: 3.5
                petal_length:
                  type: number
                  format: float
                  example: 1.4
                petal_width:
                  type: number
                  format: float
                  example: 0.2
                predicted_class:
                  type: integer
                  example: 0
                created_at:
                  type: string
                  format: date-time
                  example: 2023-10-01T12:34:56Z
        401:
          description: Token inválido ou ausente.